+++ src/components/uyghur/index.jsx
@@ -1,7 +1,10 @@
 import React, { Component } from 'react';
 
 // semantic-ui
-import { Container, Header } from 'semantic-ui-react';
+import {
+  Container,
+  Header
+} from 'semantic-ui-react';
 
 // router
 import {
@@ -11,14 +14,19 @@ import {
 } from 'react-router-dom';
 
 // pages
+import UyghurBoycott  from './boycott'
+
 import UyghurDirectory from './directory';
+
 import UyghurArticle from './education/article';
 import UyghurNowThis from './education/nowthis';
+import UyghurResources from './education/resources';
+
+import UyghurPeople from './people';
+
 import PressureCongress from './pressure/PressureCongress'
 import PressureBrands from './pressure/PressureBrands'
-import UyghurResources from './education/resources';
-import UyghurBoycott  from './boycott'
-import UyghurPeople from './people';
+
 
 class UyghurPages extends Component {
   render() {
@@ -41,10 +49,11 @@ class UyghurPages extends Component {
             <UyghurResources />
           </Route>
 
-          <Route path="/uyghur/pressure/media">
-            <Header>
+          <Route path="/uyghur/pressure/brands">
+            <Header as="h2">
+              Pressure American brands to stop using forced Uyghur labor
               <Header.Subheader>
-                Use this page to contact brands involved in the crisis.
+                Urge brands to terminate all contracts with Xinjiang manufacturers benefitting from the forced labor of Uyghurs
               </Header.Subheader>
             </Header>
             <PressureBrands />
